requirements/dev.txt
flake8==3.4.1
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.5.0
py==1.4.34
pytest==3.2.3
pluggy==0.5.2
six==1.11.0
tox==2.9.1
virtualenv==15.1.0
coverage==4.4.1
certifi==2017.7.27.1
chardet==3.0.4
coveralls==1.2.0
docopt==0.6.2
idna==2.6
requests==2.18.4
urllib3==1.22
pg8000==1.11.0
testing.common.database==2.0.2
testing.postgresql==1.3.0
astroid==1.5.3
isort==4.2.15
lazy-object-proxy==1.3.1
pylint==1.7.4
wrapt==1.10.11
